requirements.txt
# Core AI and LangGraph dependencies
langchain==0.3.26
langgraph==0.4.8
langchain-openai==0.3.24

# A2A Protocol dependencies
aiohttp==3.12.13
requests==2.32.3

# Salesforce integration
simple-salesforce==1.12.6

# Data processing and extraction
trustcall==0.0.39
pydantic==2.11.7

# Database and storage
aiosqlite==0.21.0

# Memory and graph processing
networkx==3.4.2
numpy==1.26.4
sentence-transformers==2.2.2

# UI dependencies
textual==4.0.0

# Utilities
python-dotenv==1.0.1